--- src/main/java/com/example/pawgetherbe/controller/AccountApi.java
@@ -1,186 +1,0 @@
-package com.example.pawgetherbe.controller;
-
-import com.example.pawgetherbe.common.exceptionHandler.CustomException;
-import com.example.pawgetherbe.config.OauthConfig;
-import com.example.pawgetherbe.controller.dto.UserDto.EmailCheckRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.NickNameCheckRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.OAuth2ResponseBody;
-import com.example.pawgetherbe.controller.dto.UserDto.SignInUserRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.SignInUserResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.SignInUserWithRefreshTokenResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.UserSignUpRequest;
-import com.example.pawgetherbe.mapper.UserMapper;
-import com.example.pawgetherbe.usecase.jwt.RefreshUseCase;
-import com.example.pawgetherbe.usecase.users.DeleteUserUseCase;
-import com.example.pawgetherbe.usecase.users.EditUserUseCase;
-import com.example.pawgetherbe.usecase.users.SignInUseCase;
-import com.example.pawgetherbe.usecase.users.SignOutUseCase;
-import com.example.pawgetherbe.usecase.users.SignUpWithIdUseCase;
-import com.example.pawgetherbe.usecase.users.SignUpWithOauthUseCase;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseCookie;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import java.io.IOException;
-import java.util.Map;
-
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.INVALID_FORMAT_EMAIL;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.INVALID_FORMAT_NICKNAME;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.OAUTH_PROVIDER_NOT_SUPPORTED;
-import static com.example.pawgetherbe.common.filter.JwtAuthFilter.REQUEST_HEADER_AUTH;
-import static com.example.pawgetherbe.service.UserService.REFRESH_TOKEN_VALIDITY_SECONDS;
-import static com.example.pawgetherbe.util.ValidationUtil.isValidEmail;
-import static com.example.pawgetherbe.util.ValidationUtil.isValidNickName;
-
-
-@RestController
-@RequestMapping("/api/v1/account")
-@RequiredArgsConstructor
-@Slf4j
-public class AccountApi {
-    private final SignUpWithIdUseCase signUpWithIdUseCase;
-    private final SignUpWithOauthUseCase signUpWithOauthUseCase;
-    private final DeleteUserUseCase deleteUserUseCase;
-    private final SignOutUseCase signOutUseCase;
-    private final EditUserUseCase editUserUseCase;
-    private final SignInUseCase signInUseCase;
-    private final RefreshUseCase refreshUseCase;
-
-    private final UserMapper userMapper;
-    private final OauthConfig oauthConfig;
-
-    private static final String SAME_SITE_STRICT = "Strict";
-
-    @PostMapping("/signup")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void signup(@Validated @RequestBody UserSignUpRequest signUpRequest){
-        signUpWithIdUseCase.signUp(signUpRequest);
-    }
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void logout(@CookieValue(value = "refresh_token") String refreshToken) {
-        signOutUseCase.signOut(refreshToken);
-    }
-
-    @DeleteMapping
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteAccount(@CookieValue(value = "refresh_token") String refreshToken) {
-        deleteUserUseCase.deleteAccount(refreshToken);
-    }
-
-    @PostMapping("/signup/email")
-    @ResponseStatus(HttpStatus.OK)
-    public void signupEmailCheck(@RequestBody EmailCheckRequest emailCheckRequest) {
-        var email = emailCheckRequest.email();
-        if (!isValidEmail(email)) {
-            throw new CustomException(INVALID_FORMAT_EMAIL);
-        }
-        signUpWithIdUseCase.signupEmailCheck(email);
-    }
-
-    @PostMapping("/signup/nickname")
-    @ResponseStatus(HttpStatus.OK)
-    public void signupNickNameCheck(@RequestBody NickNameCheckRequest nickNameCheckRequest){
-        var nickName = nickNameCheckRequest.nickName();
-        if (!isValidNickName(nickName)) {
-            throw new CustomException(INVALID_FORMAT_NICKNAME);
-        }
-        signUpWithIdUseCase.signupNicknameCheck(nickName);
-    }
-
-    @PatchMapping
-    public UpdateUserResponse updateUserInfo(@RequestBody UpdateUserRequest request) {
-        var updateUserResponse = editUserUseCase.updateUserInfo(request);
-
-        return updateUserResponse;
-    }
-
-    @GetMapping("/oauth/{provider}")
-    public void redirectToProvider(@PathVariable String provider, HttpServletResponse response) throws IOException {
-        var providerProps = oauthConfig.getProviders().get(provider);
-        if (providerProps == null) {
-            throw new CustomException(OAUTH_PROVIDER_NOT_SUPPORTED);
-        }
-
-        String redirectUrl = UriComponentsBuilder
-                .fromHttpUrl(providerProps.getAuthorizationUri())
-                .queryParam("client_id", providerProps.getClientId())
-                .queryParam("redirect_uri", providerProps.getRedirectUri())
-                .queryParam("response_type", "code")
-                .queryParam("scope", String.join(" ", providerProps.getScope()))
-                .build()
-                .toUriString();
-
-        response.sendRedirect(redirectUrl);
-    }
-
-    @GetMapping("/oauth/callback/{provider}")
-    public ResponseEntity<OAuth2ResponseBody> oauthCallback(
-            @PathVariable String provider,
-            @RequestParam String code) {
-        var oauth2SignUpResponse = signUpWithOauthUseCase.oauthSignUp(provider,code);
-
-        ResponseCookie cookie = buildRefreshTokenCookieHeader(oauth2SignUpResponse.refreshToken());
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, cookie.toString())
-                .body(userMapper.toOAuth2ResponseBody(oauth2SignUpResponse));
-    }
-
-    @PostMapping
-    public ResponseEntity<SignInUserResponse> signIn(@RequestBody @Valid SignInUserRequest signInRequest) {
-        SignInUserWithRefreshTokenResponse user = signInUseCase.signIn(signInRequest);
-
-        ResponseCookie refreshTokenCookieHeader = buildRefreshTokenCookieHeader(user.refreshToken());
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, refreshTokenCookieHeader.toString())
-                .body(userMapper.toSignInUserResponse(user));
-    }
-
-    @PostMapping("/refresh")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<String> refresh(@RequestHeader(REQUEST_HEADER_AUTH) String authHeader,
-                                          @CookieValue(name = "refreshToken") String refreshToken) {
-
-        Map<String, String> authTokens = refreshUseCase.refresh(authHeader, refreshToken);
-
-        ResponseCookie refreshTokenCookieHeader = buildRefreshTokenCookieHeader(authTokens.get("refreshToken"));
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, refreshTokenCookieHeader.toString())
-                .body(authTokens.get("accessToken"));
-    }
-
-    private ResponseCookie buildRefreshTokenCookieHeader(String refreshToken) {
-        return ResponseCookie.from("refreshToken", refreshToken)
-                .httpOnly(true)
-                .secure(true)
-                .path("/")
-                .maxAge(REFRESH_TOKEN_VALIDITY_SECONDS)
-                .sameSite(SAME_SITE_STRICT)
-                .build();
-    }
-}
