+++ src/main/java/com/example/pawgetherbe/controller/query/dto/UserQueryDto.java
@@ -1,0 +1,12 @@
+package com.example.pawgetherbe.controller.query.dto;
+
+public final class UserQueryDto {
+
+    public record EmailCheckRequest(
+            String email
+    ) {}
+
+    public record NickNameCheckRequest(
+            String nickName
+    ) {}
+}
