--- src/main/java/com/example/pawgetherbe/mapper/LikeMapper.java
@@ -1,7 +1,0 @@
-package com.example.pawgetherbe.mapper;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface LikeMapper {
-}
