+++ src/main/java/com/example/pawgetherbe/mapper/command/BookmarkCommandMapper.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.mapper.command;
+
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface BookmarkCommandMapper {
+}
