+++ src/main/java/com/example/pawgetherbe/mapper/command/ImageCommandMapper.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.mapper.command;
+
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface ImageCommandMapper {
+}
