+++ src/main/java/com/example/pawgetherbe/mapper/command/PetFairCommandMapper.java
@@ -1,0 +1,4 @@
+package com.example.pawgetherbe.mapper.command;
+
+public interface PetFairCommandMapper {
+}
