+++ src/main/java/com/example/pawgetherbe/mapper/query/BookmarkQueryMapper.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.mapper.query;
+
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface BookmarkQueryMapper {
+}
