+++ src/main/java/com/example/pawgetherbe/mapper/query/ReplyQueryMapper.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.mapper.query;
+
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface ReplyQueryMapper {
+}
