+++ src/main/java/com/example/pawgetherbe/mapper/query/UserQueryMapper.java
@@ -1,0 +1,9 @@
+package com.example.pawgetherbe.mapper.query;
+
+import com.example.pawgetherbe.controller.command.dto.UserCommandDto;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface UserQueryMapper {
+    UserCommandDto.UserAccessTokenDto toAccessTokenDto(Long userId, String userRole);
+}
