--- src/main/java/com/example/pawgetherbe/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.example.pawgetherbe.repository;
-
-import com.example.pawgetherbe.domain.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    boolean existsByEmail(String email);
-    UserEntity findByEmail(String email);
-
-    boolean existsByNickName(String nickName);
-
-    void deleteByEmail(String email);
-}
