+++ src/main/java/com/example/pawgetherbe/repository/command/UserCommandRepository.java
@@ -1,0 +1,12 @@
+package com.example.pawgetherbe.repository.command;
+
+import com.example.pawgetherbe.domain.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserCommandRepository extends JpaRepository<UserEntity, Long> {
+    boolean existsByEmail(String email);
+    UserEntity findByEmail(String email);
+    boolean existsByNickName(String nickName);
+}
