+++ src/main/java/com/example/pawgetherbe/repository/query/UserQueryRepository.java
@@ -1,0 +1,11 @@
+package com.example.pawgetherbe.repository.query;
+
+import com.example.pawgetherbe.domain.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserQueryRepository extends JpaRepository<UserEntity, Long> {
+    boolean existsByEmail(String email);
+    boolean existsByNickName(String nickName);
+}
