--- src/main/java/com/example/pawgetherbe/service/UserService.java
@@ -1,387 +1,0 @@
-package com.example.pawgetherbe.service;
-
-import com.example.pawgetherbe.common.exceptionHandler.CustomException;
-import com.example.pawgetherbe.common.oauth.OAuthProviderSpec;
-import com.example.pawgetherbe.config.OauthConfig;
-import com.example.pawgetherbe.controller.dto.UserDto.SignInUserRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.SignInUserWithRefreshTokenResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.UserAccessTokenDto;
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.Oauth2SignUpResponse;
-import com.example.pawgetherbe.controller.dto.UserDto.UserSignUpRequest;
-import com.example.pawgetherbe.domain.entity.OauthEntity;
-import com.example.pawgetherbe.domain.entity.UserEntity;
-import com.example.pawgetherbe.domain.status.AccessTokenStatus;
-import com.example.pawgetherbe.domain.status.UserRole;
-import com.example.pawgetherbe.domain.status.UserStatus;
-import com.example.pawgetherbe.mapper.UserMapper;
-import com.example.pawgetherbe.repository.OauthRepository;
-import com.example.pawgetherbe.repository.UserRepository;
-import com.example.pawgetherbe.usecase.jwt.RefreshUseCase;
-import com.example.pawgetherbe.usecase.users.DeleteUserUseCase;
-import com.example.pawgetherbe.usecase.users.EditUserUseCase;
-import com.example.pawgetherbe.usecase.users.SignInUseCase;
-import com.example.pawgetherbe.usecase.users.SignOutUseCase;
-import com.example.pawgetherbe.usecase.users.SignUpWithIdUseCase;
-import com.example.pawgetherbe.usecase.users.SignUpWithOauthUseCase;
-import com.example.pawgetherbe.util.EncryptUtil;
-import com.example.pawgetherbe.util.JwtUtil;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.scribejava.core.builder.ServiceBuilder;
-import com.github.scribejava.core.model.OAuth2AccessToken;
-import com.github.scribejava.core.model.OAuthRequest;
-import com.github.scribejava.core.model.Response;
-import com.github.scribejava.core.model.Verb;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.CONFLICT_EMAIL;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.CONFLICT_NICKNAME;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.CONFLICT_USER;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.NOT_FOUND_USER;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.OAUTH_PROVIDER_NOT_SUPPORTED;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.UNAUTHORIZED_LOGIN;
-import static com.example.pawgetherbe.exception.command.UserCommandErrorCode.UNAUTHORIZED_PASSWORD;
-import static com.example.pawgetherbe.common.filter.JwtAuthFilter.AUTH_BEARER;
-import static com.example.pawgetherbe.domain.UserContext.getUserId;
-import static com.example.pawgetherbe.util.EncryptUtil.generateRefreshToken;
-import static com.example.pawgetherbe.util.EncryptUtil.passwordEncode;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class UserService implements SignUpWithIdUseCase, SignUpWithOauthUseCase, DeleteUserUseCase, SignOutUseCase, EditUserUseCase,
-        SignInUseCase, RefreshUseCase {
-
-    private final UserRepository userRepository;
-    private final OauthRepository oauthRepository;
-    private final UserMapper userMapper;
-    private final RedisTemplate<String, String> redisTemplate;
-
-    private final JwtUtil jwtUtil;
-    private final OauthConfig oauthConfig;
-
-    public static final int REFRESH_TOKEN_VALIDITY_SECONDS = 7 * 24 * 60 * 60; // 7일
-
-    @Override
-    @Transactional
-    public void signUp(UserSignUpRequest request) {
-        if(userRepository.existsByEmail(request.email())){
-            throw new CustomException(CONFLICT_USER);
-        }
-        var userEntity = userMapper.toUserEntity(request);
-        var userEntityBuilder = userEntity
-                    .toBuilder()
-                    .password(passwordEncode(request.password()))
-                    .status(UserStatus.ACTIVE)
-                    .role(UserRole.USER_EMAIL)
-                    .build();
-
-        userRepository.save(userEntityBuilder);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public void signupEmailCheck(String email) {
-        if (userRepository.existsByEmail(email)) {
-            throw new CustomException(CONFLICT_EMAIL);
-        }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public void signupNicknameCheck(String nickName) {
-        if(userRepository.existsByNickName(nickName)) {
-            throw new CustomException(CONFLICT_NICKNAME);
-        }
-    }
-
-    @Override
-    @Transactional
-    public Oauth2SignUpResponse oauthSignUp(String provider, String code) {
-        log.info("oauthSignUp start");
-        Map<String, Object> userInfo = fetchOAuthUserInfo(provider, code);
-
-        String email = (String) userInfo.get("email");
-        String nickName = (String) userInfo.get("nickname");
-        String providerId = (String) userInfo.get("providerId");
-        log.info("email = {}", email);
-        log.info("nickName = {}", nickName);
-        log.info("providerId = {}", providerId);
-
-        var oauthCheck = oauthRepository.existsByOauthProviderIdAndOauthProvider(providerId, provider);
-        var userCheck = userRepository.existsByEmail(email);
-
-        UserEntity user;
-
-        if(userCheck && !oauthCheck) {
-            user = userRepository.findByEmail(email);
-            var oauthEntity = OauthEntity.builder()
-                    .oauthProviderId(providerId)
-                    .oauthProvider(provider)
-                    .user(user)
-                    .build();
-            oauthRepository.save(oauthEntity);
-            var token = getToken(user);
-            user.updateRole(UserRole.USER_BOTH);
-
-            return userMapper.toOauth2SignUpResponse(
-                    user,
-                    null,
-                    token.get("accessToken"),
-                    token.get("refreshToken"));
-        } else if (oauthCheck){
-            user = oauthRepository.findByOauthProviderId(providerId).get().getUser();
-        }else {
-            if (userRepository.existsByNickName(nickName)){
-                nickName = UUID.randomUUID().toString().substring(0, 8);
-            }
-            UserEntity newUser = UserEntity.builder()
-                    .email(email)
-                    .nickName(nickName)
-                    .password(passwordEncode(UUID.randomUUID().toString().substring(0, 8))) // 랜덤 값
-                    .status(UserStatus.ACTIVE)
-                    .role(UserRole.USER_AUTH)
-                    .build();
-
-            user = userRepository.save(newUser);
-
-            var oauthEntity = OauthEntity.builder()
-                    .oauthProviderId(providerId)
-                    .oauthProvider(provider)
-                    .user(user)
-                    .build();
-            oauthRepository.save(oauthEntity);
-        }
-
-        var token = getToken(user);
-
-        return userMapper.toOauth2SignUpResponse(
-                user,
-                provider,
-                token.get("accessToken"),
-                token.get("refreshToken")
-        );
-    }
-
-    @Override
-    @Transactional
-    public void deleteAccount(String refreshToken) {
-        var id = Long.valueOf(getUserId());
-        var oauthCheck = oauthRepository.existsByUser_Id(id);
-        if(oauthCheck) {
-            oauthRepository.deleteByUser_Id(id);
-        }
-        userRepository.deleteById(id);
-        redisTemplate.delete(refreshToken);
-    }
-
-    @Override
-    public void signOut(String refreshToken) {
-        redisTemplate.delete(refreshToken);
-    }
-
-    @Override
-    @Transactional
-    public UpdateUserResponse updateUserInfo(UpdateUserRequest request) {
-        var id = getUserId();
-        var user = userRepository.findById(Long.valueOf(id)).orElseThrow(() ->
-            new CustomException(NOT_FOUND_USER)
-        );
-
-        user.updateProfile(request.nickName(), request.userImg());
-
-        return userMapper.toUpdateUserResponse(request);
-    }
-
-    public Map<String, String> getToken(UserEntity userEntity) {
-        String refreshToken = generateRefreshToken();
-        String accessToken = jwtUtil.generateAccessToken(
-                new UserAccessTokenDto(userEntity.getId(), userEntity.getRole())
-        );
-
-        // Redis에 저장 (key: refreshToken, value: userId)
-        redisTemplate.opsForValue().set(refreshToken, String.valueOf(userEntity.getId()), Duration.ofDays(7));
-
-        // 토큰을 Map으로 반환
-        Map<String, String> tokenMap = new HashMap<>();
-        tokenMap.put("refreshToken", refreshToken);
-        tokenMap.put("accessToken", accessToken);
-        return tokenMap;
-    }
-
-    public Map<String, Object> fetchOAuthUserInfo(String provider, String code){
-        var oauth = oauthConfig.getProviders().get(provider);
-
-        if (oauth == null) {
-            throw new CustomException(OAUTH_PROVIDER_NOT_SUPPORTED);
-        }
-        log.info("getClientId = {}", oauth.getClientId());
-        log.info("clientSecret = {}", oauth.getClientSecret());
-
-//        OAuth20Service service = null;
-
-        var service = new ServiceBuilder(oauth.getClientId())
-                .apiSecret(oauth.getClientSecret())
-                .callback(oauth.getRedirectUri())
-                .defaultScope(String.join(" ", oauth.getScope()))
-                .build(new OAuthProviderSpec(
-                        oauth.getAuthorizationUri(),
-                        oauth.getTokenUri()
-                ));
-
-        try {
-            OAuth2AccessToken accessToken = service.getAccessToken(code);
-
-            OAuthRequest request = new OAuthRequest(Verb.GET, oauth.getUserInfoUri());
-            service.signRequest(accessToken, request);
-            Response response = service.execute(request);
-
-            ObjectMapper mapper = new ObjectMapper();
-            JsonNode root = mapper.readTree(response.getBody());
-
-            return switch (provider) {
-                case "google" -> Map.of(
-                        "email", root.path("email").asText(),
-                        "nickname", root.path("nickname").asText(),
-                        "providerId", root.path("sub").asText()
-                );
-                case "naver" -> {
-                    JsonNode naverResponse = root.path("response");
-                    yield Map.of(
-                            "email", naverResponse.path("email").asText(),
-                            "nickname", naverResponse.path("nickname").asText(),
-                            "providerId", naverResponse.path("id").asText()
-                    );
-                }
-                default -> throw new CustomException(OAUTH_PROVIDER_NOT_SUPPORTED);
-            };
-
-        }catch (Exception e) {
-            throw new RuntimeException("OAuth 처리 중 오류 발생", e);
-        }
-    }
-//    private Map<String, Object> fetchKakaoUserInfo(String code, Provider oauth) {
-//        // 1. 토큰 요청
-//        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-//        params.add("grant_type", "authorization_code");
-//        params.add("client_id", oauth.getClientId());
-//        params.add("redirect_uri", oauth.getRedirectUri());
-//        params.add("code", code);
-//
-//        if (StringUtils.hasText(oauth.getClientSecret())) {
-//            params.add("client_secret", oauth.getClientSecret());
-//        }
-//
-//        HttpHeaders headers = new HttpHeaders();
-//        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-//        HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(params, headers);
-//
-//        ResponseEntity<JsonNode> tokenResponse = new RestTemplate()
-//                .postForEntity(oauth.getTokenUri(), request, JsonNode.class);
-//
-//        String accessToken = tokenResponse.getBody().get("access_token").asText();
-//
-//        // 2. 사용자 정보 요청
-//        HttpHeaders userInfoHeaders = new HttpHeaders();
-//        userInfoHeaders.setBearerAuth(accessToken);
-//        HttpEntity<Void> userInfoRequest = new HttpEntity<>(userInfoHeaders);
-//
-//        ResponseEntity<JsonNode> userInfoResponse = new RestTemplate()
-//                .exchange(oauth.getUserInfoUri(), HttpMethod.GET, userInfoRequest, JsonNode.class);
-//
-//        JsonNode root = userInfoResponse.getBody();
-//        JsonNode kakaoAccount = root.path("kakao_account");
-//        JsonNode profile = kakaoAccount.path("profile");
-//
-//        String email = kakaoAccount.path("email").asText(null);
-//        if (email == null || email.isEmpty()) {
-//            email = "random-" + UUID.randomUUID().toString().substring(0, 8) + "@example.com";
-//        }
-//
-//        return Map.of(
-//                "email", email,
-//                "nickname", profile.path("nickname").asText(),
-//                "providerId", root.path("id").asText()
-//        );
-//    }
-
-    @Override
-    @Transactional
-    public SignInUserWithRefreshTokenResponse signIn(SignInUserRequest signInRequest) {
-
-        UserEntity userEntity = userRepository.findByEmail(signInRequest.email());
-
-        // 회원 존재 유무 체크
-        if (userEntity == null) {
-            throw new CustomException(NOT_FOUND_USER);
-        }
-
-        // password 유효성 확인
-        if (!isMatchedPassword(signInRequest.password(), userEntity.getPassword())) {
-            throw new CustomException(UNAUTHORIZED_PASSWORD);
-        }
-
-        // access token 발급
-        String accessToken = jwtUtil.generateAccessToken(new UserAccessTokenDto(userEntity.getId(), userEntity.getRole()));
-
-        // refresh token 발급
-        String refreshToken = EncryptUtil.generateRefreshToken();
-        redisTemplate.opsForValue().set(refreshToken, String.valueOf(userEntity.getId()), Duration.ofDays(REFRESH_TOKEN_VALIDITY_SECONDS));
-
-        return userMapper.toSignInWithRefreshToken(userEntity, accessToken, refreshToken);
-    }
-
-    @Override
-    public Map<String, String> refresh(String authHeader, String refreshToken) {
-
-        if (!authHeader.startsWith(AUTH_BEARER)) {
-            throw new CustomException(UNAUTHORIZED_LOGIN);
-        }
-
-        String accessToken = authHeader.substring(AUTH_BEARER.length());
-
-        if (jwtUtil.validateToken(accessToken).equals(AccessTokenStatus.INVALID)) {
-            throw new CustomException(UNAUTHORIZED_LOGIN);
-        }
-
-        // case1] refresh token 만료: 재로그인
-        if (!isValidRefreshToken(refreshToken)) {
-            throw new CustomException(UNAUTHORIZED_LOGIN);
-        }
-
-        Long userId = jwtUtil.getUserIdFromToken(accessToken);
-        String userRole = jwtUtil.getUserRoleFromToken(accessToken);
-
-        // case2] refresh token 만료 X: 갱신 로직
-        UserEntity user= userRepository.findById(userId).orElseThrow(
-                () -> new CustomException(NOT_FOUND_USER)
-        );
-
-        String renewAccessToken = jwtUtil.generateAccessToken(userMapper.toAccessTokenDto(userId, userRole));
-        String renewRefreshToken = EncryptUtil.generateRefreshToken();
-
-        return Map.of(
-                "accessToken", renewAccessToken,
-                "refreshToken", renewRefreshToken
-        );
-    }
-
-    private boolean isMatchedPassword(String plainPassword, String encryptPassword) {
-        return EncryptUtil.matches(plainPassword, encryptPassword);
-    }
-
-    private boolean isValidRefreshToken(String refreshToken) {
-        return redisTemplate.opsForValue().get(refreshToken) != null;
-    }
-}
