--- src/main/java/com/example/pawgetherbe/usecase/jwt/IssueAccessTokenUseCase.java
@@ -1,4 +1,0 @@
-package com.example.pawgetherbe.usecase.jwt;
-
-public interface IssueAccessTokenUseCase {
-}
