--- src/main/java/com/example/pawgetherbe/usecase/jwt/RefreshUseCase.java
@@ -1,7 +1,0 @@
-package com.example.pawgetherbe.usecase.jwt;
-
-import java.util.Map;
-
-public interface RefreshUseCase {
-    Map<String, String> refresh(String authHeader, String refreshToken);
-}
