+++ src/main/java/com/example/pawgetherbe/usecase/jwt/query/RefreshQueryUseCase.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.usecase.jwt.query;
+
+import java.util.Map;
+
+public interface RefreshQueryUseCase {
+    Map<String, String> refresh(String authHeader, String refreshToken);
+}
