--- src/main/java/com/example/pawgetherbe/usecase/users/DeleteUserUseCase.java
@@ -1,5 +1,0 @@
-package com.example.pawgetherbe.usecase.users;
-
-public interface DeleteUserUseCase {
-    void deleteAccount(String refreshToken);
-}
