--- src/main/java/com/example/pawgetherbe/usecase/users/EditUserUseCase.java
@@ -1,8 +1,0 @@
-package com.example.pawgetherbe.usecase.users;
-
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserRequest;
-import com.example.pawgetherbe.controller.dto.UserDto.UpdateUserResponse;
-
-public interface EditUserUseCase {
-    UpdateUserResponse updateUserInfo(UpdateUserRequest request);
-}
