--- src/main/java/com/example/pawgetherbe/usecase/users/SignOutUseCase.java
@@ -1,5 +1,0 @@
-package com.example.pawgetherbe.usecase.users;
-
-public interface SignOutUseCase {
-    void signOut(String refreshToken);
-}
