--- src/main/java/com/example/pawgetherbe/usecase/users/SignUpWithIdUseCase.java
@@ -1,9 +1,0 @@
-package com.example.pawgetherbe.usecase.users;
-
-import com.example.pawgetherbe.controller.dto.UserDto;
-
-public interface SignUpWithIdUseCase {
-    void signUp(UserDto.UserSignUpRequest request);
-    void signupEmailCheck(String email);
-    void signupNicknameCheck(String nickname);
-}
