--- src/main/java/com/example/pawgetherbe/usecase/users/SignUpWithOauthUseCase.java
@@ -1,7 +1,0 @@
-package com.example.pawgetherbe.usecase.users;
-
-import com.example.pawgetherbe.controller.dto.UserDto.Oauth2SignUpResponse;
-
-public interface SignUpWithOauthUseCase {
-    Oauth2SignUpResponse oauthSignUp(String provider, String code);
-}
