+++ src/main/java/com/example/pawgetherbe/usecase/users/command/DeleteUserCommandUseCase.java
@@ -1,0 +1,5 @@
+package com.example.pawgetherbe.usecase.users.command;
+
+public interface DeleteUserCommandUseCase {
+    void deleteAccount(String refreshToken);
+}
