+++ src/main/java/com/example/pawgetherbe/usecase/users/command/EditUserCommandUseCase.java
@@ -1,0 +1,8 @@
+package com.example.pawgetherbe.usecase.users.command;
+
+import com.example.pawgetherbe.controller.command.dto.UserCommandDto.UpdateUserRequest;
+import com.example.pawgetherbe.controller.command.dto.UserCommandDto.UpdateUserResponse;
+
+public interface EditUserCommandUseCase {
+    UpdateUserResponse updateUserInfo(UpdateUserRequest request);
+}
