+++ src/main/java/com/example/pawgetherbe/usecase/users/command/SignOutCommandUseCase.java
@@ -1,0 +1,5 @@
+package com.example.pawgetherbe.usecase.users.command;
+
+public interface SignOutCommandUseCase {
+    void signOut(String refreshToken);
+}
