+++ src/main/java/com/example/pawgetherbe/usecase/users/command/SignUpCommandOauthUseCase.java
@@ -1,0 +1,7 @@
+package com.example.pawgetherbe.usecase.users.command;
+
+import com.example.pawgetherbe.controller.command.dto.UserCommandDto.Oauth2SignUpResponse;
+
+public interface SignUpCommandOauthUseCase {
+    Oauth2SignUpResponse oauthSignUp(String provider, String code);
+}
