+++ src/main/java/com/example/pawgetherbe/usecase/users/query/SignUpQueryUseCase.java
@@ -1,0 +1,6 @@
+package com.example.pawgetherbe.usecase.users.query;
+
+public interface SignUpQueryUseCase {
+    void signupEmailCheck(String email);
+    void signupNicknameCheck(String nickname);
+}
